rename: layout 컴포넌트를 template 폴더로 이동
