+++ pages/_app.tsx
@@ -1,7 +1,7 @@
 import '../styles/globals.css';
 import type { AppProps } from 'next/app';
 import { RecoilRoot } from 'recoil';
-import Layout from 'components/Layout';
+import Layout from 'components/template/Layout';
 
 function App({ Component, pageProps }: AppProps) {
   return (
